feature/v2.8.0
支持Imap协议
升级所有依赖
修复部分bug

+++ server/controllers/email/delete.go
@@ -11,8 +11,8 @@ import (
 )
 
 type emailDeleteRequest struct {
-	IDs       []int64 `json:"ids"`
-	ForcedDel bool    `json:"forcedDel"`
+	IDs       []int `json:"ids"`
+	ForcedDel bool  `json:"forcedDel"`
 }
 
 func EmailDelete(ctx *context.Context, w http.ResponseWriter, req *http.Request) {
